+++ Gruntfile.js
@@ -1,8 +1,8 @@
-var ServiceNow = require('./services/snclient.js');
 var require_config = require("./helper/config_validator");
 
 var fs = require('fs');
 var path = require('path');
+
 module.exports = function (grunt) {
 	require('load-grunt-tasks')(grunt);
 
@@ -38,57 +38,7 @@ module.exports = function (grunt) {
 //    grunt.loadNpmTasks('grunt-contrib-watch');
 //    grunt.task.renameTask("watch", "watchAndPush");
 //	grunt.registerTask("pullSN",[]);
-	grunt.registerTask('pullSN', 'Pull command.', function (folderName) {
-        
-		var done = this.async();
-        require_config().then(function (config) {
-			
-			var snHelper = new ServiceNow(config);
-
-			var query = "name" + "STARTSWITH" + 'solution';
-
-			
-			
-			var folder_path = path.join(destination,folderName);
-
-			if(!fs.existsSync(folder_path)){
-				fs.mkdirSync(folder_path);
-			}
-			
-			snHelper.getRecords(config.folders[folderName].table,query).on("complete",function(data,response){
-				if(data.code === 'ECONNREFUSED' ){
-					console.log(data);
-				} 
-				else{
-					for(var i = 0; i < data.result.length; i++){
-						var content = data.result[i][config.folders[folderName].field];
-						var filename = data.result[i].name;
-
-						if('extension' in config.folders[folderName]){
-							filename = filename + "." + snConfig.folders[folderName].extension;
-						}
-						var file_path = path.join(folder_path, filename);
-						fs.writeFile(file_path, content, function(err){
-
-							if(err){
-								console.error("ERR",err);
-							}
-							else{
-								console.log("Touching file " + file_path);
-								// why is this showing only the second files name twice
-
-							}
-							if(i === data.result.length){
-								done();
-							}
-						});
-					}
-				}
-				done();
-				
-			});
-		});
-	});
+	grunt.task.loadTasks('./tasks');
 	
 
 };
